requirements.txt
openai==1.30.1
python-dotenv==1.0.1
langchain==0.2.0
langchain-community==0.2.0
langchain-openai==0.1.7
tiktoken==0.7.0